Create star_classification.py
A fun project where I try to predict from data if something is a star, galaxy, or a quasar. I use a similar pipeline where I can test many models and many different hyperparameters to test for the best model. MSE, MAE, and R2 are used to determine the best model and the best model results are shown on screen.